feat: async event emitter
